docker-images/controller/requirements.txt
redis
GitPython
git+https://oauth2:53gZwasv6UmExxrohqPm@gitlab.com/steven.horn/covid-prevalence.git